+++ lib/src/main/kotlin/app/cash/quiver/extensions/List.kt
@@ -1,7 +1,9 @@
 package app.cash.quiver.extensions
 
+import arrow.core.Either
 import arrow.core.Option
 import arrow.core.filterOption
+import arrow.core.raise.either
 
 /**
  * Returns a list without Nones