+++ lib/src/test/kotlin/app/cash/quiver/extensions/EitherTest.kt
@@ -11,15 +11,20 @@ import io.kotest.assertions.arrow.core.shouldBeLeft
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.assertions.arrow.core.shouldBeSome
 import io.kotest.core.spec.style.StringSpec
+import io.kotest.matchers.result.shouldBeFailure
+import io.kotest.matchers.result.shouldBeSuccess
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.int
 import io.kotest.property.arrow.core.either
 import io.kotest.property.checkAll
 import io.kotest.property.arbitrary.list
+import io.kotest.property.arbitrary.map
+import io.kotest.property.arbitrary.string
 import io.kotest.property.arrow.core.option
 import org.junit.jupiter.api.assertThrows
 import app.cash.quiver.extensions.traverse as quiverTraverse
+import app.cash.quiver.extensions.traverseOption as quiverTraverseOption
 import app.cash.quiver.extensions.sequence as quiverSequence
 
 class EitherTest : StringSpec({
@@ -120,48 +125,93 @@ class EitherTest : StringSpec({
       res shouldBe expected
     }
   }
+
   "traverse should return transformed list of Right when Right" {
     checkAll(Arb.int(), Arb.list(Arb.int())) { a, bs ->
       Either.Right(a).quiverTraverse { bs } shouldBe bs.map { it.right() }
     }
   }
+
   "singleton list when Left" {
     checkAll(Arb.int(), Arb.list(Arb.int())) { a, bs ->
       Either.Left(a).quiverTraverse { bs } shouldBe listOf(Either.Left(a))
     }
   }
+
   "sequence should return list of Right when all are Right in list" {
     checkAll(Arb.list(Arb.int())) { bs ->
       bs.right().quiverSequence() shouldBe bs.map { it.right() }
     }
   }
+
   "sequence should return first left" {
     checkAll(Arb.int()) { a ->
       val aa: Either<Int, List<String>> = a.left()
       aa.quiverSequence() shouldBe listOf(a.left())
     }
   }
+
   "traverse should return transformed option as Right when Right" {
     checkAll(Arb.int(), Arb.option(Arb.int())) { a, option ->
       Either.Right(a).quiverTraverse { option } shouldBe option.map { it.right() }
     }
   }
-  "option of Left when Left" {
+
+  "traverse should return option of Left when Left" {
     checkAll(Arb.int(), Arb.option(Arb.int())) { a, option ->
       Either.Left(a).quiverTraverse { option } shouldBe Some(Either.Left(a))
     }
   }
+
+  "traverseOption should return transformed option as Right when Right" {
+    checkAll(Arb.int(), Arb.option(Arb.int())) { a, option ->
+      Either.Right(a).quiverTraverseOption { option } shouldBe option.map { it.right() }
+    }
+  }
+
+  "traverseOption should return Option of Left when Left" {
+    checkAll(Arb.int(), Arb.option(Arb.int())) { a, option ->
+      Either.Left(a).quiverTraverseOption { option } shouldBe Some(Either.Left(a))
+    }
+  }
+
   "sequence should return list of Some when all are Some in list" {
     checkAll(Arb.option(Arb.int())) { a ->
       a.right().quiverSequence() shouldBe a.map { it.right() }
     }
   }
+
   "sequence should return first None" {
     checkAll(Arb.int()) { a ->
       val aa: Either<Int, Option<String>> = a.left()
       aa.quiverSequence() shouldBe a.left().some()
     }
   }
+
+  "errorOr can always be a result" {
+    checkAll(Arb.either(
+      left = Arb.string().map { Exception(it) },
+      right = Arb.int()
+    )) { either ->
+      either.toResult().fold(
+        onFailure = { either.shouldBeLeft() shouldBe it },
+        onSuccess = { either.shouldBeRight() shouldBe it },
+      )
+    }
+  }
+
+  "handleFailureWith recovers errors" {
+    1.right().handleErrorWith { Either.Left(Throwable("never happens")) }.shouldBeRight(1)
+    Throwable("sad panda").left().handleErrorWith { Throwable("even sadder").left() }
+      .shouldBeLeft().message shouldBe "even sadder"
+
+    Throwable("sad panda").left().handleErrorWith { 1.right() }
+      .shouldBeRight(1)
+
+    Throwable("sad panda").left().handleErrorWith { it.left() }
+      .shouldBeLeft().message shouldBe "sad panda"
+  }
+
 })
 
 fun testParse(s: String): ErrorOr<Int> =